fix(landing): 데모 아이콘 폭을 짝수로, 화살표에 대체 텍스트 부여 (리뷰 #2 #3)
DemoChrome 은 devup 린트 규칙(파일명=컴포넌트명)에 맞춰
DemoHeading.tsx / DemoArrow.tsx 로 분리했다.
