+++ apps/landing/src/app/DemoHeading.tsx
@@ -1,4 +1,4 @@
-import { Box, Flex, Image, Text } from '@devup-ui/react'
+import { Box, Flex, Text } from '@devup-ui/react'
 
 /** 데모 섹션 상단의 손가락 아이콘 + 안내 문구. 한글·수학 데모가 공유한다. */
 export function DemoHeading({ children }: { children: string }) {
@@ -17,34 +17,13 @@ export function DemoHeading({ children }: { children: string }) {
         maskPosition="center"
         maskRepeat="no-repeat"
         maskSize="contain"
-        w={['17px', null, null, '28px']}
+        // 아이콘 원본 비율(28x32)상 h=20px 의 이론 폭은 17.5px 이다.
+        // maskSize=contain 이라 박스를 18px 로 잡아도 아이콘은 늘어나지 않는다.
+        w={['18px', null, null, '28px']}
       />
       <Text color="$text" pos="relative" top="-2px" typography="mainTextSm">
         {children}
       </Text>
-    </Flex>
-  )
-}
-
-/** 입력 박스와 출력 박스 사이의 방향 화살표. 한글·수학 데모가 공유한다. */
-export function DemoArrow() {
-  return (
-    <Flex aria-hidden="true">
-      <Image
-        alt=""
-        display={['none', null, null, 'block']}
-        mr="10px"
-        role="presentation"
-        src="/images/home/translate-arrow-circle.svg"
-        w="16px"
-      />
-      <Image
-        alt=""
-        role="presentation"
-        src="/images/home/translate-arrow.svg"
-        transform={[null, null, null, 'rotate(-90deg)']}
-        w={['16px', null, null, '24px']}
-      />
     </Flex>
   )
 }
